requirements.txt
absl-py==1.1.0
apiflask==1.1.0
apispec==5.2.2
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cmake==3.22.5
colorama==0.4.5
cycler==0.11.0
deepface
face-recognition-models==0.3.0
filelock==3.7.1
flask==2.1.2
flask-httpauth==4.7.0
flask-marshmallow==0.14.0
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
imutils==0.5.4
jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
keras-preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
markdown==3.3.7
markupsafe==2.1.1
marshmallow==3.17.0
matplotlib==3.5.2
mtcnn-opencv==1.0.2
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.10
waitress==2.1.2
webargs==8.1.0
werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0